requirements.txt
atlassian-python-api==3.20.1
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
oauthlib==3.2.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
smmap==5.0.0
typed-config==1.1.0
typing_extensions==4.1.1
urllib3==1.26.8
wrapt==1.13.3
